+++ src/Time/DeltaT/DeltaTInterface.php
@@ -1,22 +1,15 @@
 <?php
-declare(strict_types = 1);
 
-/**
- * Interface for Delta T calculation
- *
- * @author    Marcus Jaschen <mjaschen@gmail.com>
- * @license   http://www.opensource.org/licenses/mit-license MIT License
- * @link      https://www.marcusjaschen.de/
- */
+declare(strict_types=1);
 
 namespace Astrotools\Time\DeltaT;
 
 /**
- * Interface for Delta T calculation
+ * Interface for Delta T calculation.
  *
  * @author   Marcus Jaschen <mail@marcusjaschen.de>
  * @license  http://www.opensource.org/licenses/mit-license MIT License
- * @link     https://www.marcusjaschen.de/
+ * @see     https://www.marcusjaschen.de/
  */
 interface DeltaTInterface
 {